requirements.txt
beautifulsoup4
persist-queue
python-slugify
requests
urllib3